requirements.txt
fastapi
gunicorn
uvicorn
jinja2
pytest
requests
pre-commit
python-dotenv

# uploading related
aiofiles
python-multipart
pillow

pytesseract